refactor: integrate Clock component functionality directly into Home page for improved structure and remove unused Clock file and tests
