+++ src/app/__tests__/page.a11y.test.tsx
@@ -5,11 +5,11 @@ expect.extend(matchers);
 import React from 'react';
 import { render } from '@testing-library/react';
 import { axe } from 'vitest-axe';
-import Clock from '../Clock';
+import Home from '../page';
 
-describe('App accessibility', () => {
-  it('main Clock page has no accessibility violations', async () => {
-    const { container } = render(<Clock />);
+describe('Page accessibility', () => {
+  it('main page has no accessibility violations', async () => {
+    const { container } = render(<Home />);
     // @ts-expect-error: Custom matcher from vitest-axe
     expect(await axe(container)).toHaveNoViolations();
   });
